Update patch set 3
Patch Set 3: Verified+1

Looks pretty good. There are a few screens/buttons/menus that still could use an upgrade but this is potentially enough to start.

One issue though is the Settings Activity: the heading titles are White and hard to read. Also, the checkboxes are bright white and hard to see when selected.

Patch-set: 3