Update patch set 16
Patch Set 16:

Danny you were right about the text alignment it was due to gravity.  

Regarding the padding, I think it is simpler to specify the padding on the message bubble rather than each item inside the bubble.  That way it isn't possible to forget the padding on one of the items, which is why they were not lined up.

I made the horizontal and vertical padding separate since that seems to be what you were going for, although I wouldn't mind making them symmetrical also.

Patch-set: 16