Update patch set 4
Patch Set 4:

This was working for us on one msim device that we tried. There may have been some differences though, so I'd like to fix the crash that you're seeing. Please post a log here or email it to me: matt@cyngn.com. MultiSim might not be totally ready on CM12 yet, I'm not sure as to the state of it yet. So, an unrelated bug might be causing the problems you're seeing. I will try to reproduce your issues on another msim device today.

Patch-set: 4