Update patch set 9
Patch Set 9:

We should work into having the preferences set to SIM cards instead of slots (and use the SIM card names instead), but for now I'd just like to see the "SIM1" and "SIM2" strings to be "SIM 1" and "SIM 2" to be more consistent with the rest of the places where we already say that...

Also, can you please rebase this patch?

Thanks.

Patch-set: 9