Update patch set 1
Patch Set 1:

Yes, it will break long sender names. But if alignBaseline is moved to "date", how shall I refer to "from" which has not been created yet? I tried android:layout_alignBaseline="@+id/from" in "date" and android:id="@id/from" in "from". This is working as expected, but it looks a bit messy. I'll upload this patchset in a few minutes.

Patch-set: 1