Update patch set 1
Patch Set 1: Code-Review-1

We probably should remove the pattern pref in code as well on those devices though and not let it dangle around.

The point of the dependency attribute is different, though: It makes sure the dependant pref is disabled if the master pref is unchecked. How is that ensured on devices with vibrator after your patch?

Patch-set: 1