Update patch set 2
Patch Set 2:

I'd like for that to be case, however the caf/12.0 branch is upstream for some caf projects and they don't segregate strings vs caf_strings. So they won't take this change and we'll be out of sync for some of our projects. Hoping to remedy this in the upcoming iterations (12.1?)

Patch-set: 2