Update patch set 1
Patch Set 1:

No,I dont talk about this .of course ,we can type replace button to replace attachment.
for example
we have attachment like pic,type to add another attachment,the button color of Vcal in Attachment box turn to gray( mean to disable),but it can be typed.
so i just let it diable

Patch-set: 1