Update patch set 3
Patch Set 3:

Pengfei, when I commented saying that the comment needs to be < 100 char, I meant that it has to be less than <100 char *per line* not that the whole comment block has to be less than one line. You just have to line wrap when it goes over the 100 char limit. This isn't a huge deal, it's just a style guide that's enforced across the projects for CM just to keep code readable and styled the same way as AOSP.

I just tested your patch and I can't seem to reproduce the "Can't save draft after adding a contact attachment" bug. And I'm not seeing any code in your patch that would affect this. The select recipients bug does get fixed in your patch so that checks out.

Patch-set: 3