Update patch set 3
Patch Set 3:

The code looks good to me for what you're doing, changing phoneId to subId. I'm not familiar enough with the Telephony stack to approve this change though.

Where is the best place in the code to see documentation for how Phone ID vs Subscription ID are defined? I'm not completely sure of the difference, although I can make an educated guess that they mean the physical slot (phone ID) vs an ID number associated with the SIM itself (subscription ID)?

Patch-set: 3