Update patch set 3
Patch Set 3:

The phone ID is the index of the phone instance in the telephony manager (as per getPhone(int)), which at this point matches the slot index, even though it isn't the absolutely same thing. CAF uses it interchangeably though, so correcting that would be a larger refactoring. The subscription ID is the _id value of the respective SIM's roe in the SIM info DB table.

Patch-set: 3