Update patch set 7
Patch Set 7:

I doubt it's the initial size parameter. ContentValues are backed by a HashMap, which definitely grows once it has exceeded the initial size. Maybe there still is something else wonky, so we better should trace down the path of the sub ID.

Tony, for you it was broken before this patch as well, right? Can you give me a log of a failed send attempt?

Patch-set: 7