Update patch set 7
Patch Set 7:

Ok, I'm getting this as the mmsUri in WorkingMessage.java function: sendMmsWorker - mmsUri: content://mms/drafts/278

Then most of the rest of my logging in there doesn't get hit, so it appears the message is stored as a draft.  Maybe that's why it's bypassing the db inserts. I'm looking for the route a draft takes vs the way I've been looking.

Patch-set: 7