Update patch set 8
Patch Set 8: Code-Review+1

Danny, Does PS 8 look ok.  There was no real reason to move things below the p.move(), because if I duplicated the code, it'd run in either scenario, so I just reversed the if and left both scenarios doing the db work.

So far so good on my tests!

Patch-set: 8