Update patch set 8
Patch Set 8:

Danny, also notice the comments I removed.  They refer to that trigger you'd pointed to earlier.  So in this case, in this PS, we're not relying on that trigger, it's more of a forced move.

Not sure, but something doesn't seem correct about this.  It seems to work, but that doesn't mean it's right.

Patch-set: 8