Update patch set 8
Patch Set 8:

Yeah, my only thing is... Everything would have worked fine if all we did was add the phone_id column back to the db in telephonyprovider. 

Google messenger would use that as needed, and mms-caf would not since it didn't care to.

Maybe I'm just being naive, but a lot of times the simplest solution can be the best.

Not saying I'm right, just saying.

Patch-set: 8