Update patch set 8
Patch Set 8:

Ok on sending:  We got a null APN, then the wrong APN.  After the timeout, it tried the next APN in the list, and that APN was right.  Once the correct APN was selected the message sent.      It appears that the receive never works because there is no "retry".  It fails to download, then gives you the "Download" button.  When you hit that, it recreates the connection for MMS from the start, getting the wrong APN again. So it never downloads.

Patch-set: 8