Update patch set 9
Patch Set 9:

With the new patch, the sub_id is definitely populated fine for me: http://pastebin.com/HJgBCdLV

Please give it a try. If it still doesn't work for you, please post some steps on how exactly you create the MMS.

BTW, semi-OT: I just realized that it's kind of counterintuitive to use the _data_ subscription for MMS in the MSIM case: When I select 'Always ask' as SMS subscription, I get the prompt for MMS as well, but when I select any SIM for SMS, it won't necessarily use that. Any thoughts on using the default SMS sub in both cases, maybe naming the option in settings accordingly?

Patch-set: 9