Update patch set 10
Patch Set 10:

It's looking better, but on recieve, there's a ClassCastException converting integer to long from Bundle.  So I'm guessing there's at least one more place where a long is put into extras but an int is expected, or vice versa.

If I find it, I'll push a patch.

Patch-set: 10