Update patch set 13
Patch Set 13: Code-Review+1

Got a couple, seemingly innocent, errors in the log.  Other than that, it's looking good.

Here's the errors if you wanna see: http://pastebin.com/Pdv3rewN

Mms and sms seem to send and receive just fine, but those are in the logs, so it's worth mentioning.

Patch-set: 13