Update patch set 13
Patch Set 13:

further note, it appears that part of the log is created by TransactionSettings.java in the constructor.

If the String apnName is passed it, it tries to use that and I get the first message with the wrong apn and that reference to looked for = null, but if there is no apnname passed in, it logs the looked for apn=?.

So it looks like initially something is passing in the wrong apn.

Patch-set: 13