Update patch set 13
Patch Set 13:

I think it does add a delay, and if you look through the logs, it gives a PDU timeout error, and a retry.  I'm not as worried about the delay as I am the initial failure/timeout. In the end it may work, but I don't think we should have to go through an error for it to work.

Patch-set: 13