Update patch set 13
Patch Set 13:

That didn't help, but looking at what's happening on TransactionSettings.java, starting at line 90, it querries the carriers database.

This returns basically all the apn's you'd see in settings, and there's no filter I can see, nor think of that'll be able to select the right apn the first time.  So when apnName goes into here as null, it does it's best guess against the database, returning the first record which is pretty much always wrong.

Patch-set: 13