Update patch set 13
Patch Set 13:

Ok, but for me, it does return an mmsc.  It's just not the right apn.  I'm on t-mobile, and I get the metropcs apn selected that first time.

It works coming from onMmsPduConnected(), where it get the apnName from NetworkInfo instead.  But the first mms send attempt appears to not use that method and just uses the one that sends a null apnName.  I've querried the db directly and don't see anything unique in the carriers table that could be searched on in this case.  As of right now, I think we're either lucky it goes to that pdu method, or that's by design and I haven't traced it that route yet.

Patch-set: 13