Update patch set 1
Patch Set 1:

I'll see if I can get that from the user.

I do have a thought on what might be going on for him though.

He has two apns: Optus Internet and Optus MMS,  In apns-conf.xml, only Optus MMS is set as type mms.  So the only way I see that the "internet" apn could be passed into transactionsettings, is by onMmsPdpConnected in TransactionService.

I'm not sure on where that gets called from originally, but it does pass in the apnName to TransactionSettings.

I propose using that apnName in TransactionService via mmsNetworkInfo.getExtraInfo(), to query the apn type and verify that it allows mms.  If it does, proceed, if it does not, call the new TransactionSettings leaving apnName variable passed in as null.  Thus forcing TransactionSettings to query for the correct apn on it's own and not rely on the invalid one being passed in.

Does that sound reasonable?

Patch-set: 1