Update patch set 1
Patch Set 1: -Code-Review

"if it does not, call the new TransactionSettings leaving apnName variable passed in as null. "

"if it does not" will meet mms sending failure.In a mms  sending progress there are only two steps,
first, query apns  with apnName=null,and got a apn used to create pdp connection(get a ip and so on...)
second, query apns with apnName=xxx(which gained in step 1) and used to got the MMSC. If MMSC is not empty, will try to connect and send data. if MMSC is empty will use check next apn.(this logic in TransactionSettings())

Patch-set: 1