Update patch set 3
Patch Set 3:

(7 comments)

This patch still seems kind of broken for me. When I do repro steps in the commit message, the contact that is still visible after mobile numbers only is re-checked, still shows as checked. but when I press back, the contact doesn't get added to the recipients editor.

Patch-set: 3