Update patch set 1
Patch Set 1:

If that's the case, makes more sense to use a cm_strings entry for this and re-wire the source to point to that entry. Looks like there are only a handful of uses, should be a small enough change, and we can get more appropriate translations that way.

Patch-set: 1