Update patch set 1
Patch Set 1:

Probably, I would have chosen a Handler.Callback to send the Editable as part of the message, instead of have it as a local var. With the current code, the Edibale could be replaced while the thread is running. Passing it as a local var you won't have that problem.

Patch-set: 1