Update patch set 8
Patch Set 8:

@Joey : The attachment editor will change to look like : http://www.cyanogenmod.org/blog/messaging-redesign/attachment/1-0-0-in-message-attachments-center-page. 

We are working on rolling out these changes in stages and modifying the attachment editor wasn't slated for this sprint. The goal is to get a functioning 'new app' as quickly as possible and adding features incrementally.

I will pass along adding fav contacts to FAB feedback to Aaron (the designer owning the project). 

Also, feel free to submit patches or take on the implementation of any of the features that seem interesting.

Patch-set: 8