Update patch set 19
Patch Set 19:

Since the jira ticket that this references is on the cyngn internal tracker I can't find the answer out on my own. Has there been any work on resolving the issue where Mms-caf reproducably displays the wrong contact picture for contacts sometimes (seems to be a cache issue or something)?

Patch-set: 19