Update patch set 1
Patch Set 1:

As it is, I see messages in the status bar as 'John DoeHello, how are you?' instead of 'John Doe: Hello, how are you?'.  After some more thought last night, I believe maybe your intent was to remove the ': ' if the subject and body are simultaneously empty.  That would require a !(a && b) instead of the (!a && !b ) that was originally there.  If that was your intent, I'll update this patch.

Patch-set: 1