Update patch set 2
Patch Set 2:

It matches the color scheme of the app, so it's fine from that viewpoint. The actual icon shape looks good to me as well.

I wonder whether it's possible to talk you guys into a different, more neutral, color scheme though ... what was wrong about the cyan we chose before?

Patch-set: 2