Update patch set 1
Patch Set 1:

There are vestiges of support in QuickMessage to reply to multiple messages.

Aaron is undecided on whether to fix support for multiple messages or to always have QuickMessage apply only to the last conversation. Leaving the ViewPager / multi-message support in case we need to go back to it

Patch-set: 1