Update patch set 2
Patch Set 2:

heads up /notification should also be disabled when popup is on. Heads up overlaps with popup when there's an incoming message. And when i make a new sms and i type the name of a contact i want to send the message to, if that has 2 or more numbers i'm just shown the "category" and not the number itself (ie: i see Joey Rizzoli Home and Work and not the 2 numbers), it should be better to show them too imho

Patch-set: 2