Update patch set 7
Patch Set 7:

I think there is a simpler way to approach this problem. In onPrepareActionMode(), look at the checked item count from the list view and if it is greater than 0, then follow the same actions as you would in onItemCheckedStateChanged. This will for cases where you in the activity and initiate action mode and also when you come back.

This avoids us having to track a lot of state. 

The ListView itself is persisting state information b/w restarts etc. Just take advantage of that rather than having to implement these additional measures.

Patch-set: 7