Update patch set 1
Patch Set 1:

Ok, so it looks like we take care of the default setting starting out as false, then removing the capability to automagically enable itself.  This seems to work fine for a fresh install.  How do we want to handle the setting in an upgrade scenario?  This will only effect new devices.

Patch-set: 1