Update patch set 1
Patch Set 1: -Code-Review

Not sure if we are going to make this conversations instead of messages.  It seems to make more sense to have unread conversations rather than unread messages.  How should we handle mark unread conversations?  If we have 20 conversations with 1000 messages each, this could be a _long_ running task

Patch-set: 1